backend/requirements.txt
amqp==5.0.9
APScheduler==3.8.1
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.12b0
blinker==1.4
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.3.2
dbus-python==1.2.16
Deprecated==1.2.13
distro==1.5.0
Django==4.0.1
django-apscheduler==0.6.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
docutils==0.16
et-xmlfile==1.1.0
h11==0.12.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.0.1
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
linecache2==1.0.0
louis==3.18.0
macaroonbakery==1.3.1
more-itertools==4.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.15
netifaces==0.10.9
numpy==1.22.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.1.3
paramiko==2.7.2
pathspec==0.9.0
pbr==5.6.0
pexpect==4.8.0
Pillow==8.1.2
platformdirs==2.4.1
pop-transition==1.1.2
prompt-toolkit==3.0.24
protobuf==3.12.4
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
Pygments==2.7.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyodbc==4.0.30
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
python-apt===2.2.1pop0-1628521831-21.10-5dbe6cf
python-dateutil==2.8.2
python-debian===0.1.39ubuntu1
python-dotenv==0.19.2
python-xlib==0.29
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.3.1
redis==4.1.1
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
roman==2.0.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.1.0
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spotipy==2.19.0
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
tomli==1.2.3
traceback2==1.4.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.2
unittest2==1.1.0
urllib3==1.26.5
vine==5.0.0
wadllib==1.3.5
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
zipp==1.0.0
